:new: Resume slime job upon preemption and show on dashboard
